Fixed error in pipeline script that failed to print an error message to the Pipeline_error.log if job status is KILLED, STOPPED, FAILED, ARCHIVING_FAILED (i.e. fatal to completion).
Note there is a bug in the Agave API that sometimes inserts a FAILED status in a job that is not failed (causing my code to make the wrong decision about a job), but at this point I have not removed code that will act on a FAILED status and stop looking for output. That is, I am assuming this bug in the API will be fixed soom as I have been in touch with the devs.

Also with this commit we are removing the file /xGDBvm/admin/xgdbpass which is a dynamic flag that should not have been in the repository.